2-3단계 피드백 반영 (#649)
* 글, 댓글에 작성자 이름 주입방식 변경

* 로그인과 권한 확인을 서로 분리

* PathPattern 정리

* UserListShow를 controller에서만 사용하도록 수정

+++ src/main/java/com/kakao/cafe/web/ReplyController.java
@@ -1,13 +1,13 @@
 package com.kakao.cafe.web;
 
 import com.kakao.cafe.domain.entity.User;
+import com.kakao.cafe.dto.reply.ReplyContents;
 import com.kakao.cafe.dto.reply.ReplyCreateCommand;
 import com.kakao.cafe.service.ReplyService;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.SessionAttribute;
+import org.springframework.web.bind.annotation.*;
+
+import javax.naming.NoPermissionException;
 
 @Controller
 public class ReplyController {
@@ -17,17 +17,36 @@ public class ReplyController {
         this.replyService = replyService;
     }
 
-    @PostMapping("/questions/{articleId}/answers/")
+    @RequestMapping(value = "/questions/{articleId}/answers/", method = RequestMethod.POST)
     public String writeReply(@PathVariable long articleId,
-                             @SessionAttribute("sessionedUser") User user,
+                             @RequestAttribute("sessionedUser") User user,
                              String contents) {
         replyService.createReply(new ReplyCreateCommand(articleId, user.getUserId(), contents));
         return "redirect:/articles/{articleId}";
     }
 
-    @DeleteMapping("/questions/{articleId}/answers/{replyId}")
-    public String deleteReply(@PathVariable long replyId) {
-        replyService.deleteReply(replyId);
-        return "redirect:/articles/{articleId}";
+    @RequestMapping(value = "/questions/{articleId}/answers/{replyId}",
+            method = RequestMethod.DELETE)
+    public String deleteReply(@PathVariable long replyId,
+                              @RequestAttribute("sessionedUser") User user) {
+        try {
+            checkPermission(user, replyId);
+            replyService.deleteReply(replyId);
+            return "redirect:/articles/{articleId}";
+        } catch (NoPermissionException e) {
+            return "redirect:/nopermission";
+        }
+    }
+
+    private ReplyContents checkPermission(User user, long replyId) throws NoPermissionException{
+        ReplyContents replyContents = replyService.getReply(replyId);
+        if (userNotPermittedToReply(user, replyContents)) {
+            throw new NoPermissionException();
+        }
+        return replyContents;
+    }
+
+    private boolean userNotPermittedToReply(User user, ReplyContents replyContents) {
+        return !user.getUserId().equals(replyContents.getWriterId());
     }
 }
